Linking to Sphinx HTML as Standalone Files
By setting `markdown_http_base` to the root of a Sphinx website and
`markdown_target_ext` to `.html`, it's possible to write a document that
can be used both within Sphinx-generated HTML and as separate standalone
markdown file, which will link to the Sphinx-generated HTML.

This was added to create GitHub release notes from a Sphinx/RST source.

Also added `markdown_ext` to control output extension, defaults to `.md`
`markdown_target_ext` will default to `markdown_ext` if not set.